Rename authorization file
